Add compulsory prefix "tmp_" to temporary filename
Fixes Issue #86.

The call to `File.createTempFile` causes an `IllegalArgumentException` as it is expecting the filename parameter to be > 3 characters. Therefore a fatal error occurs when filename prefixes are less than 3 characters e.g.: 3.jpg, ab.jpg etc.

To fix this I have added a compulsory prefix to the created temporary file.